+++ tests/unit/transformer/test_mapper.py
@@ -2,23 +2,28 @@ from builtins import str
 from builtins import object
 from typing import List
 from copy import deepcopy
+from uuid import uuid4
 
 import pytest
 from pyspark.sql import functions as F
 from pyspark.sql import types as T
 from pyspark.sql import Row
+from pyspark.sql import DataFrame
+from pyspark.sql import SparkSession
 from pyspark.sql.utils import AnalysisException
 from chispa.dataframe_comparer import assert_df_equality
 
-from spooq.transformer.mapper import DataTypeValidationFailed
+from spooq.transformer.annotator import Annotator, AnnotatorMode
+from spooq.transformer.mapper import ColumnMappingNotSupported, DataTypeValidationFailed, MapperMode
 from tests import DATA_FOLDER
 from spooq.transformer import Mapper
+from spooq.transformer.mapper import MapperMode, MissingColumnHandling
 from spooq.transformer import mapper_transformations as spq
 
 
 @pytest.fixture(scope="module")
 def transformer(mapping):
-    return Mapper(mapping=mapping, missing_column_handling="nullify")
+    return Mapper(mapping=mapping, missing_column_handling=MissingColumnHandling.nullify)
 
 
 @pytest.fixture(scope="module")
@@ -98,6 +103,28 @@ class TestBasicAttributes(object):
         assert str(transformer) == "Transformer Object of Class Mapper"
 
 
+class TestLenghtOfMappingTuple:
+    @pytest.fixture(scope="class")
+    def input_df(self, spark_session: SparkSession):
+        return spark_session.range(0)
+
+    def test_column_mapping_without_comments(self, input_df: DataFrame):
+        mapping = [("id", "id", T.LongType())]
+        output_df = Mapper(mapping).transform(input_df)
+        assert_df_equality(input_df, output_df)
+
+    def test_column_mapping_with_comments(self, input_df: DataFrame):
+        mapping = [("id", "id", T.LongType(), "This column contains the ID")]
+        output_df = Mapper(mapping).transform(input_df)
+        assert_df_equality(input_df, output_df, ignore_metadata=True)
+
+    @pytest.mark.parametrize("num_of_elements", (0, 1, 2, 5, 6))
+    def test_unsupported_number_of_elements_for_column_mapping(self, input_df: DataFrame, num_of_elements: int):
+        mapping = ["string" for x in range(0, num_of_elements)]
+        with pytest.raises(ColumnMappingNotSupported):
+            Mapper(mapping).transform(input_df)
+
+
 class TestShapeOfMappedDataFrame(object):
     def test_same_amount_of_records(self, input_df, mapped_df):
         """Amount of Rows is the same after the transformation"""
@@ -142,16 +169,16 @@ class TestMultipleMappings(object):
 
     def test_appending_a_mapping(self, mapped_df, new_mapping, input_columns, new_columns):
         """Output schema is correct for added mapping at the end of the input schema"""
-        new_mapped_df = Mapper(mapping=new_mapping, mode="append", missing_column_handling="nullify").transform(
-            mapped_df
-        )
+        new_mapped_df = Mapper(
+            mapping=new_mapping, mode=MapperMode.append, missing_column_handling=MissingColumnHandling.nullify
+        ).transform(mapped_df)
         assert input_columns + new_columns == new_mapped_df.columns
 
     def test_prepending_a_mapping(self, mapped_df, new_mapping, input_columns, new_columns):
         """Output schema is correct for added mapping at the beginning of the input schema"""
-        new_mapped_df = Mapper(mapping=new_mapping, mode="prepend", missing_column_handling="nullify").transform(
-            mapped_df
-        )
+        new_mapped_df = Mapper(
+            mapping=new_mapping, mode=MapperMode.prepend, missing_column_handling=MissingColumnHandling.nullify
+        ).transform(mapped_df)
         assert new_columns + input_columns == new_mapped_df.columns
 
     def test_appending_a_mapping_with_duplicated_columns(self, input_columns, mapped_df):
@@ -163,9 +190,9 @@ class TestMultipleMappings(object):
         ]
         new_columns = [name for (name, path, data_type) in new_mapping]
         new_columns_deduplicated = [x for x in new_columns if x not in input_columns]
-        new_mapped_df = Mapper(mapping=new_mapping, mode="append", missing_column_handling="nullify").transform(
-            mapped_df
-        )
+        new_mapped_df = Mapper(
+            mapping=new_mapping, mode=MapperMode.append, missing_column_handling=MissingColumnHandling.nullify
+        ).transform(mapped_df)
         assert input_columns + new_columns_deduplicated == new_mapped_df.columns
         assert mapped_df.schema["birthday"].dataType == T.TimestampType()
         assert new_mapped_df.schema["birthday"].dataType == T.DateType()
@@ -179,9 +206,9 @@ class TestMultipleMappings(object):
         ]
         new_columns = [name for (name, path, data_type) in new_mapping]
         new_columns_deduplicated = [x for x in new_columns if x not in input_columns]
-        new_mapped_df = Mapper(mapping=new_mapping, mode="prepend", missing_column_handling="nullify").transform(
-            mapped_df
-        )
+        new_mapped_df = Mapper(
+            mapping=new_mapping, mode=MapperMode.prepend, missing_column_handling=MissingColumnHandling.nullify
+        ).transform(mapped_df)
         assert new_columns_deduplicated + input_columns == new_mapped_df.columns
         assert mapped_df.schema["birthday"].dataType == T.TimestampType()
         assert new_mapped_df.schema["birthday"].dataType == T.DateType()
@@ -194,7 +221,7 @@ class TestExceptionForMissingInputColumns(object):
 
     @pytest.fixture(scope="class")
     def transformer(self, mapping):
-        return Mapper(mapping=mapping, missing_column_handling="raise_error")
+        return Mapper(mapping=mapping, missing_column_handling=MissingColumnHandling.raise_error)
 
     def test_missing_column_raises_exception(self, input_df, transformer):
         input_df = input_df.drop("attributes")
@@ -214,7 +241,7 @@ class TestNullifyMissingColumns(object):
 
     @pytest.fixture(scope="class")
     def transformer(self, mapping):
-        return Mapper(mapping=mapping, missing_column_handling="nullify")
+        return Mapper(mapping=mapping, missing_column_handling=MissingColumnHandling.nullify)
 
     @pytest.fixture(scope="class")
     def mapped_df(self, input_df, transformer):
@@ -241,7 +268,7 @@ class TestSkipMissingColumns(object):
 
     @pytest.fixture(scope="class")
     def transformer(self, mapping):
-        return Mapper(mapping=mapping, missing_column_handling="skip")
+        return Mapper(mapping=mapping, missing_column_handling=MissingColumnHandling.skip)
 
     @pytest.fixture(scope="class")
     def mapped_df(self, input_df, transformer):
@@ -342,7 +369,7 @@ class TestValidateDataTypes:
         assert_df_equality(input_df, transformer.transform(input_df))
 
     def test_matching_mapping_with_validation(self, input_df, matching_mapping):
-        transformer = Mapper(matching_mapping, mode="rename_and_validate")
+        transformer = Mapper(matching_mapping, mode=MapperMode.rename_and_validate)
         assert_df_equality(input_df, transformer.transform(input_df))
 
     def test_mismatching_mapping_with_casting(self, input_df, matching_mapping):
@@ -356,14 +383,14 @@ class TestValidateDataTypes:
     def test_mismatching_mapping_with_validation(self, input_df, matching_mapping):
         mapping_ = deepcopy(matching_mapping)
         mapping_[0] = ("col_a", "col_a", T.StringType())
-        transformer = Mapper(mapping_, mode="rename_and_validate")
+        transformer = Mapper(mapping_, mode=MapperMode.rename_and_validate)
         with pytest.raises(DataTypeValidationFailed, match="col_a"):
             transformer.transform(input_df)
 
     def test_spooq_transformation_raises_exception(self, input_df, matching_mapping):
         mapping_ = deepcopy(matching_mapping)
         mapping_[0] = ("col_a", "col_a", spq.to_int)
-        transformer = Mapper(mapping_, mode="rename_and_validate")
+        transformer = Mapper(mapping_, mode=MapperMode.rename_and_validate)
         with pytest.raises(
             DataTypeValidationFailed, match="Spooq transformations are not allowed in 'rename_and_validate' mode"
         ):
@@ -373,13 +400,15 @@ class TestValidateDataTypes:
         missing_column = "col_d"
         mapping_ = deepcopy(matching_mapping)
         mapping_.append((missing_column, missing_column, T.StringType()))
-        transformer = Mapper(mapping_, mode="rename_and_validate", missing_column_handling="raise_error")
+        transformer = Mapper(
+            mapping_, mode=MapperMode.rename_and_validate, missing_column_handling=MissingColumnHandling.raise_error
+        )
         with pytest.raises(
             AnalysisException,
             match=f"A column or function parameter with name `{missing_column}` cannot be resolved.|"
-                  f"cannot resolve '{missing_column}' given input columns|"
-                  f"cannot resolve '`{missing_column}`' given input columns|"
-                  f"Column '{missing_column}' does not exist. Did you mean one of the following?",
+            f"cannot resolve '{missing_column}' given input columns|"
+            f"cannot resolve '`{missing_column}`' given input columns|"
+            f"Column '{missing_column}' does not exist. Did you mean one of the following?",
         ):
             transformer.transform(input_df)
 
@@ -387,7 +416,9 @@ class TestValidateDataTypes:
         skipped_column = "col_d"
         mapping_ = deepcopy(matching_mapping)
         mapping_.append((skipped_column, skipped_column, T.StringType()))
-        transformer = Mapper(mapping_, mode="rename_and_validate", missing_column_handling="skip")
+        transformer = Mapper(
+            mapping_, mode=MapperMode.rename_and_validate, missing_column_handling=MissingColumnHandling.skip
+        )
         with pytest.raises(DataTypeValidationFailed, match=f"Input column: {skipped_column} not found!"):
             transformer.transform(input_df)
 
@@ -395,17 +426,199 @@ class TestValidateDataTypes:
         column_to_nullify = "col_d"
         mapping_ = deepcopy(matching_mapping)
         mapping_.append((column_to_nullify, column_to_nullify, T.StringType()))
-        mapped_df = Mapper(mapping_, mode="rename_and_validate", missing_column_handling="nullify").transform(input_df)
+        mapped_df = Mapper(
+            mapping_, mode=MapperMode.rename_and_validate, missing_column_handling=MissingColumnHandling.nullify
+        ).transform(input_df)
         assert mapped_df.schema[column_to_nullify].jsonValue()["type"] == "string"
 
     def test_spooq_transformation_as_is(self, input_df, matching_mapping):
         mapping_ = deepcopy(matching_mapping)
         mapping_[0] = ("col_a", "col_a", spq.as_is)
-        transformer = Mapper(mapping_, mode="rename_and_validate")
+        transformer = Mapper(mapping_, mode=MapperMode.rename_and_validate)
         assert_df_equality(input_df, transformer.transform(input_df))
 
     def test_spooq_transformation_as_is_string(self, input_df, matching_mapping):
         mapping_ = deepcopy(matching_mapping)
         mapping_[0] = ("col_a", "col_a", "as_is")
-        transformer = Mapper(mapping_, mode="rename_and_validate")
+        transformer = Mapper(mapping_, mode=MapperMode.rename_and_validate)
         assert_df_equality(input_df, transformer.transform(input_df))
+
+
+class TestColumnComments:
+
+
+    @pytest.fixture(scope="class")
+    def setup_database(self, spark_session: SparkSession):
+        spark_session.sql("CREATE DATABASE IF NOT EXISTS db")
+        yield
+        spark_session.sql("DROP DATABASE IF EXISTS db CASCADE")
+
+    @pytest.fixture(scope="module")
+    def input_schema(self) -> str:
+        return """
+            col_a int COMMENT 'initial',
+            col_b int,
+            col_c int
+        """
+
+    @pytest.fixture(scope="class")
+    def input_df(self, spark_session: SparkSession, input_schema: str) -> DataFrame:
+        return spark_session.createDataFrame([Row(col_a=1, col_b=2, col_c=3)], input_schema)
+
+    @pytest.fixture()
+    def input_table(self, spark_session: SparkSession, input_df: DataFrame, setup_database) -> str:
+        input_df.write.saveAsTable(name="db.table_with_partial_comments", format="delta", mode="overwrite")
+        yield "db.table_with_partial_comments"
+        spark_session.sql("DROP TABLE db.table_with_partial_comments")
+
+    @pytest.fixture(scope="class")
+    def mapping(self) -> List[tuple]:
+        # fmt: off
+        return [
+            ("col_a", "col_a", spq.as_is, "updated"),
+            ("col_b", "col_b", spq.as_is, "updated"),
+            ("col_c", "col_c", spq.as_is),
+        ]
+        # fmt: on
+
+    @pytest.fixture()
+    def random_string(self) -> str:
+        return str(uuid4())[:8]
+
+    @pytest.mark.parametrize(
+        argnames=["column_name", "expected_comment"],
+        argvalues=[
+            ("col_a", "initial"),  # taken from referenced sql_source_table
+            ("col_b", None),
+            ("col_c", None),
+        ],
+    )
+    def test_fetch_comments_from_input_table(
+        self,
+        spark_session: SparkSession,
+        column_name: str,
+        expected_comment: str,
+        mapping: List,
+        input_table: str,
+        random_string: str,
+    ):
+        mapping = [
+            ("col_a", "col_a", spq.as_is),
+            ("col_b", "col_b", spq.as_is),
+            ("col_c", "col_c", spq.as_is),
+        ]
+        source_df = spark_session.table(input_table)
+        output_df = Mapper(mapping=mapping, annotator_options={"sql_source_table_identifier": input_table}).transform(
+            source_df
+        )
+        output_df.write.saveAsTable(f"db.output_table_{random_string}")
+        table_description_df = spark_session.sql(f"DESCRIBE db.output_table_{random_string}")
+        assert (
+            table_description_df.where(F.col("col_name") == column_name).rdd.map(lambda row: row.comment).collect()[0]
+            == expected_comment
+        )
+
+    @pytest.mark.parametrize(
+        argnames=["column_name", "expected_comment"],
+        argvalues=[
+            ("col_a", "initial"),  # taken from referenced sql_source_table
+            ("col_x", None),
+            ("col_y", "updated"),  # taken from mapping
+        ],
+    )
+    def test_fetch_comments_from_related_table(
+        self, spark_session: SparkSession, column_name: str, expected_comment: str, input_table: str, random_string: str
+    ):
+        source_df = spark_session.createDataFrame([], "col_a int, col_x int, col_y int")
+        mapping = [
+            ("col_a", "col_a", spq.as_is),
+            ("col_x", "col_x", spq.as_is),
+            ("col_y", "col_y", spq.as_is, "updated"),
+        ]
+        output_df = Mapper(mapping, annotator_options={"sql_source_table_identifier": input_table}).transform(source_df)
+        output_df.write.saveAsTable(f"db.output_table_{random_string}")
+        table_description_df = spark_session.sql(f"DESCRIBE db.output_table_{random_string}")
+        assert (
+            table_description_df.where(F.col("col_name") == column_name).rdd.map(lambda row: row.comment).collect()[0]
+            == expected_comment
+        )
+
+    @pytest.mark.parametrize(
+        argnames=["column_name", "expected_comment"],
+        argvalues=[
+            ("col_a", "initial"),
+            ("col_b", "updated"),
+            ("col_c", None),
+        ],
+    )
+    def test_insert_comments_to_input_table(
+        self,
+        spark_session: SparkSession,
+        column_name: str,
+        expected_comment: str,
+        mapping: List,
+        input_table: str,
+        random_string: str,
+    ):
+        source_df = spark_session.table(input_table)
+        output_df = Mapper(
+            mapping, annotator_options={"mode": AnnotatorMode.insert, "sql_source_table_identifier": input_table}
+        ).transform(source_df)
+        output_df.write.saveAsTable(f"db.output_table_{random_string}")
+        table_description_df = spark_session.sql(f"DESCRIBE db.output_table_{random_string}")
+        assert (
+            table_description_df.where(F.col("col_name") == column_name).rdd.map(lambda row: row.comment).collect()[0]
+            == expected_comment
+        )
+
+    @pytest.mark.parametrize(
+        argnames=["column_name", "expected_comment"],
+        argvalues=[
+            ("col_a", "updated"),
+            ("col_b", "updated"),
+            ("col_c", None),
+        ],
+    )
+    def test_upsert_comments_to_input_table(
+        self,
+        spark_session: SparkSession,
+        column_name: str,
+        expected_comment: str,
+        mapping: List,
+        input_table: str,
+        random_string: str,
+    ):
+        source_df = spark_session.table(input_table)
+        output_df = Mapper(mapping, annotator_options={"mode": AnnotatorMode.upsert}).transform(source_df)
+        output_df.write.saveAsTable(f"db.output_table_{random_string}")
+        table_description_df = spark_session.sql(f"DESCRIBE db.output_table_{random_string}")
+        assert (
+            table_description_df.where(F.col("col_name") == column_name).rdd.map(lambda row: row.comment).collect()[0]
+            == expected_comment
+        )
+
+    @pytest.mark.parametrize(
+        argnames=["column_name", "expected_comment"],
+        argvalues=[
+            ("col_a", "updated"),
+            ("col_b", "updated"),
+            ("col_c", None),
+        ],
+    )
+    def test_upsert_comments_to_input_table_in_validation_mode(
+        self,
+        spark_session: SparkSession,
+        column_name: str,
+        expected_comment: str,
+        mapping: List,
+        input_table: str,
+        random_string: str,
+    ):
+        source_df = spark_session.table(input_table)
+        output_df = Mapper(mapping, mode=MapperMode.rename_and_validate, annotator_options={"mode": AnnotatorMode.upsert}).transform(source_df)
+        output_df.write.saveAsTable(f"db.output_table_{random_string}")
+        table_description_df = spark_session.sql(f"DESCRIBE db.output_table_{random_string}")
+        assert (
+            table_description_df.where(F.col("col_name") == column_name).rdd.map(lambda row: row.comment).collect()[0]
+            == expected_comment
+        )